Refactor: Organize Chatbot Files into mychatbot Directory
Moves all chatbot-related HTML, CSS, and JS files into a dedicated 'mychatbot' directory.
Renames files within this directory for better clarity and consistency (e.g., chatbot-landingpage.html, chatbot-interface.js).
Consolidates two chatbot CSS files into a single chatbot-interface.css.
Updates all internal references within the chatbot files and all external references from other parts of the application (index.html, service-worker.js, documentation) to point to the new file locations and names.
