+++ index.html
@@ -29,12 +29,12 @@
   <!-- Base Styles -->
   <link rel="stylesheet" href="css/base/global.css">
   <link rel="stylesheet" href="css/base/small-screens.css">
-  <link rel="stylesheet" href="css/base/iframe-chat-wrapper.css">
+  <link rel="stylesheet" href="mychatbot/chatbot-fixed-widget.css">
 
   <!-- Modals -->
   <link rel="stylesheet" href="css/modals/contact_us_modal.css">
   <link rel="stylesheet" href="css/modals/join_us_modal.css">
-  <link rel="stylesheet" href="css/modals/chatbot_modal.css">
+  <link rel="stylesheet" href="mychatbot/chatbot-modal.css">
   <link rel="stylesheet" href="css/modals/business_operations_modal.css">
   <link rel="stylesheet" href="css/modals/contact_center_modal.css">
   <link rel="stylesheet" href="css/modals/it_support_modal.css">
@@ -320,6 +320,6 @@
   <script type="module" src="js/pages/main.js" defer></script>
   <script type="module" src="js/pages/contact_us.js" defer></script>
   <script type="module" src="js/pages/join_us.js" defer></script>
-  <script type="module" src="js/pages/chatbot.js" defer></script>
+  <script type="module" src="mychatbot/chatbot-modal.js" defer></script>
 </body>
 </html>
