+++ mychatbot/chatbot-interface.css
@@ -1,3 +1,61 @@
+/* START CONTENT FROM css/chatbot_creation/chatbot-container.css */
+/* css/chatbot_creation/chatbot-container.css */
+/* Styles for the main container of the chatbot inside the iframe */
+
+/* Base Reset and Layout */
+body {
+  margin: 0;
+  padding: env(safe-area-inset-top, 0) env(safe-area-inset-right, 0) env(safe-area-inset-bottom, 0) env(safe-area-inset-left, 0);
+  font-family: Arial, sans-serif;
+  display: flex;
+  flex-direction: column;
+  height: 100vh;
+  background-color: #f0f0f0;
+  color: #333;
+  transition: background-color 0.3s ease, color 0.3s ease;
+  line-height: 1.5;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+}
+
+/* Dark Theme */
+body[data-theme="dark"] {
+  background-color: #222;
+  color: #eee;
+}
+
+/* Main Chatbot Container */
+#chatbot-container {
+  display: flex;
+  flex-direction: column;
+  flex-grow: 1;
+  overflow: hidden;
+  margin: 2px;
+  position: relative; /* Added for positioning close button */
+  border: 2px solid rebeccapurple;
+  border-radius: 8px;
+  background-color: var(--bg-current, #fff);
+  color: var(--text-current, #333);
+  transition: border-color 0.3s ease;
+}
+
+body[data-theme="dark"] #chatbot-container {
+  border-color: mediumpurple;
+  background-color: var(--bg-current, #111);
+  color: var(--text-current, #eee);
+}
+
+/* Optional: Responsive enhancements for extremely narrow screens */
+@media (max-width: 400px) {
+  #chatbot-container {
+    margin: 0;
+    border-width: 1px;
+    border-radius: 0;
+  }
+}
+/* END CONTENT FROM css/chatbot_creation/chatbot-container.css */
+
+/* START CONTENT FROM css/chatbot_creation/chatbot-ui.css */
 /* css/chatbot_creation/chatbot-ui.css */
 /* Styles for the chat log, messages, input, button, and human check */
 
@@ -214,3 +272,4 @@ body[data-theme="dark"] #chat-send-button:hover {
      justify-content: center; /* Center checkbox content */
   }
 }
+/* END CONTENT FROM css/chatbot_creation/chatbot-ui.css */
