+++ mychatbot/chatbot-landingpage.html
@@ -23,8 +23,7 @@
   <link rel="canonical" href="https://www.opsonlinesupport.com/html/chattia.html" />
 
   <!-- Styles -->
-  <link rel="stylesheet" href="../../css/chatbot_creation/chatbot-container.css">
-  <link rel="stylesheet" href="../../css/chatbot_creation/chatbot-ui.css">
+  <link rel="stylesheet" href="./chatbot-interface.css">
 
   <!-- Google reCAPTCHA -->
   <script src="https://www.google.com/recaptcha/api.js?render=YOUR_SITE_KEY"></script>
@@ -71,6 +70,6 @@
   </div>
 
   <!-- Main JS -->
-  <script type="module" src="../../js/chatbot_creation/chatbot.js" defer></script>
+  <script type="module" src="./chatbot-interface.js" defer></script>
 </body>
 </html>
